test: merged and renamed split test files to conform with repo conventions
