+++ packages/core/src/plugins/plugin-loader.test.ts
@@ -1,10 +1,39 @@
 import { describe, it, expect, vi } from "vitest";
+import path from "node:path";
+import fs from "node:fs";
+import os from "node:os";
+
 import {
+  loadPlugins,
   PluginRuntimeRegistry,
   upsertPluginsIntoDb,
   type LoadedPlugin,
-} from "../plugins/plugin-loader";
+} from "../plugins/plugin-loader.js";
 import type { DbClient } from "../db/db.js";
+
+describe("plugin-loader", () => {
+  it("loads capabilities from a plugin directory", async () => {
+    const base = await fs.promises.mkdtemp(path.join(os.tmpdir(), "lf-plugins-"));
+    try {
+      const tmpDir = path.join(base, "fake");
+      await fs.promises.mkdir(tmpDir, { recursive: true });
+      const modContent = [
+        "module.exports = {",
+        "  capabilities: [",
+        "    { kind: 'TRIGGER', key: 'cron_schedule', displayName: 'Cron Schedule' }",
+        "  ]",
+        "};",
+      ].join("\n");
+      await fs.promises.writeFile(path.join(tmpDir, "index.js"), modContent, "utf8");
+
+      const plugins = await loadPlugins(base);
+      expect(plugins.length).toBeGreaterThan(0);
+      expect(plugins[0].capabilities[0]).toMatchObject({ key: "cron_schedule", kind: "TRIGGER" });
+    } finally {
+      await fs.promises.rm(base, { recursive: true, force: true });
+    }
+  });
+});
 
 function createFakeDb() {
   type PluginRow = { id: string; name: string };
